requirements.txt
fastapi
uvicorn
sqlalchemy
requests
pydantic>=2.9.0
yfinance
apscheduler
jinja2
GoogleNews
google-generativeai
python-dotenv